WIP move class registry to core
