+++ artichoke-backend/src/block.rs
@@ -3,8 +3,7 @@ use std::convert::TryFrom;
 use std::error;
 use std::fmt;
 
-use crate::class_registry::ClassRegistry;
-use crate::core::{TryConvertMut, Value as _};
+use crate::core::{ClassRegistry, TryConvertMut, Value as _};
 use crate::error::{Error, RubyException};
 use crate::exception_handler;
 use crate::extn::core::exception::{Fatal, TypeError};
