+++ artichoke-backend/src/extn/prelude.rs
@@ -16,7 +16,6 @@
 
 pub use crate::block::Block;
 pub use crate::class;
-pub use crate::class_registry::ClassRegistry;
 pub use crate::convert::{BoxUnboxVmValue, HeapAllocatedData};
 pub use crate::core::{Regexp as _, Value as _, *};
 pub use crate::def::{self, EnclosingRubyScope, NotDefinedError};
