+++ artichoke-backend/src/parser.rs
@@ -3,8 +3,7 @@
 
 use std::borrow::Cow;
 
-use crate::class_registry::ClassRegistry;
-use crate::core::{IncrementLinenoError, Parser, TryConvertMut};
+use crate::core::{ClassRegistry, IncrementLinenoError, Parser, TryConvertMut};
 use crate::error::{Error, RubyException};
 use crate::extn::core::exception::ScriptError;
 use crate::ffi::InterpreterExtractError;
