+++ artichoke-backend/src/string.rs
@@ -11,8 +11,7 @@ use std::borrow::{Borrow, BorrowMut, Cow};
 use std::error;
 use std::fmt;
 
-use crate::class_registry::ClassRegistry;
-use crate::core::TryConvertMut;
+use crate::core::{ClassRegistry, TryConvertMut};
 use crate::error::{Error, RubyException};
 use crate::extn::core::exception::Fatal;
 use crate::sys;
