+++ artichoke-backend/src/test/prelude.rs
@@ -12,14 +12,12 @@
 
 pub use crate::block::Block;
 pub use crate::class;
-pub use crate::class_registry::ClassRegistry;
 pub use crate::convert::{BoxUnboxVmValue, HeapAllocatedData};
 pub use crate::core::{Regexp as _, Value as _, *};
 pub use crate::def::{self, EnclosingRubyScope, NotDefinedError};
 pub use crate::error;
 pub use crate::ffi::InterpreterExtractError;
 pub use crate::module;
-pub use crate::module_registry::ModuleRegistry;
 pub use crate::prelude::*;
 pub use crate::state::parser::Context;
 pub use crate::string::{format_unicode_debug_into, WriteError};
